PD#118637: support android M

+++ DroidLogicTvInput/jni/DTVEpgScanner.c
@@ -502,7 +502,7 @@ static void epg_destroy(JNIEnv* env, jobject obj)
 {
 	EPGData *data;
 
-	data = (EPGData*)((*env)->GetLongField(env, obj, gHandleID));
+	data = (EPGData*)(long)((*env)->GetLongField(env, obj, gHandleID));
 
 	/*反注册EIT通知事件*/
 	if (data) {
@@ -527,7 +527,7 @@ static void epg_change_mode(JNIEnv* env, jobject obj, jint op, jint mode)
 	EPGData *data;
 	AM_ErrorCode_t ret;
 
-	data = (EPGData*)((*env)->GetLongField(env, obj, gHandleID));
+	data = (EPGData*)(long)((*env)->GetLongField(env, obj, gHandleID));
 
 	ret = AM_EPG_ChangeMode(data->handle, op, mode);
 	if (ret != AM_SUCCESS)
@@ -606,7 +606,7 @@ static void epg_monitor_service(JNIEnv* env, jobject obj, jobject channel)
 		ts_id = srv_id = -1;
 	}
 
-	data = (EPGData*)((*env)->GetLongField(env, obj, gHandleID));
+	data = (EPGData*)(long)((*env)->GetLongField(env, obj, gHandleID));
 	ret = AM_EPG_MonitorServiceByID(data->handle, ts_id, srv_id, PMT_Update);
 	if (ret != AM_SUCCESS) {
 		log_error("AM_EPG_MonitorService failed");
